셀러리 업데이트
