+++ backend/hey_apple_app/urls.py
@@ -7,9 +7,9 @@ urlpatterns = [
     path('fruits/<int:id>', views.get_fruit, name='fruits'),
 
     # image IO
-    path('orders', views.get_order_bill, name='orders'),
+    #path('orders', views.get_order_bill, name='orders'),
     path('orders/tasks', views.get_task_id),
 
-    path('orders', views.get_order_bill, name='orders'),
+    #path('orders', views.get_order_bill, name='orders'),
     path('bills', views.send_email_api),
 ]
